backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pandas==2.2.3
scikit-learn==1.6.1
numpy==2.2.4
joblib==1.4.2
python-multipart==0.0.20
pydantic==2.11.1
